fix react hook call issues
