+++ front/src/components/HomeComponent/ChatMessage.js
@@ -2,6 +2,6 @@ import React from "react";
 
 export default function ChatMessage(){
     return <div>
-        ChatMessage
+        {/* ChatMessage */}
     </div>
 }
